requirements.txt
arrow==1.3.0
attrs==25.3.0
caldav==2.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
icalendar==6.3.1
ics==0.7.2
idna==3.10
lxml==6.0.0
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
recurring-ical-events==3.8.0
requests==2.32.4
six==1.17.0
TatSu==5.13.1
types-python-dateutil==2.9.0.20250708
tzdata==2025.2
urllib3==2.5.0
vobject==0.9.9
x-wr-timezone==2.0.1
